A friendlier way to read tabbed auto complete for file paths
It will read the list of things completed in the terminal with "selected" after the one that is currently selected.
It assumes the command config has been completed and the last line of output is the command. From that we can gather what is selected
and add the 'selected' keyword to the string that is read aloud. This is a bit fragile but it's a start

Example command for oh my zsh is:
```
tabbing = ^(\(.*\))?➜\s{2}.+✗?\s
```