feat: Maven test result reader
Like the pytest couterpart, this parses maven output and reads the
number of failures/success, along with any errors.

Due to maven including the test count multiple times we break as soon as
we find the test count line and finish gathering errors. The output
format is a bit hard to detect as not all lines start with [error] but
hopefully it's good enough