feat: write last command output to a file
This is useful when there's a lot of output from a command and it's harder to navigate with TDSR itself.

It uses with_logging and lets TDSR just read the output so it's almost transparent. If we returned a single thing that said 'command written' we'd log that rather than the output so we return whatever we were given