feat: terraform integration
Picks out some of the key terraform lines and reads them.

The pipe character from terraform is not a normal pipe (at least on Mac) so we don't need to escape it in the regex.

There's also what looks like a blank line between errors but it's actually a upper half pipe followed by a lower half pipe.

Readme had a few typos and added some generic info to the top too to make it easy to understand what you need to do (I hope!)